requirements.txt
altair==4.2.0
async-generator==1.10
attrs==22.1.0
autobahn==19.11.2
Automat==22.10.0
bcrypt==4.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.4
decorator==5.1.1
entrypoints==0.4
enum34==1.1.10
exceptiongroup==1.0.4
extra-streamlit-components==0.1.56
gitdb==4.0.10
GitPython==3.1.29
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
Jinja2==3.1.2
jsonschema==4.17.3
kiteconnect==4.1.0
MarkupSafe==2.1.1
numpy==1.23.5
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pandas-ta==0.3.14b0
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyOpenSSL==22.1.0
pyotp==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rich==12.6.0
selenium==4.2.0
semver==2.13.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.15.2
streamlit-authenticator==0.2.1
termcolor==2.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
urllib3-secure-extra==0.1.0
validators==0.20.0
webdriver-manager==3.8.5
wsproto==1.2.0
zipp==3.11.0
zope.interface==5.5.2
